+++ server/out/index.js
@@ -51,7 +51,7 @@ function write(path, data, encoding) {
     return new Promise(resolve => {
         fs_1.default.writeFile(path, data, { encoding: "utf8" }, err => {
             if (err) {
-                // console.log("err: could not find path: ",path);
+                console.log("err: could not find path: ", path);
                 resolve(false);
             }
             else
@@ -235,6 +235,7 @@ class Story {
         let o1 = o.boards.find((v) => v != null);
         let root = new Board(s, o1.x, o1.y, o1.title, o1.text, o1.tags);
         root.img = o1.img;
+        root.audio = o1.audio;
         root._id = o1._id;
         s.start = root;
         root.load();
@@ -245,6 +246,7 @@ class Story {
                 continue;
             let b1 = new Board(s, b.x, b.y, b.title, b.text, b.tags);
             b1.img = b.img;
+            b1.audio = b.audio;
             b1._id = b._id;
             list.push(b1);
         }
@@ -386,6 +388,7 @@ class Board extends StoryObj {
     l_title;
     l_tag;
     img;
+    audio;
     load() {
         super.load();
         this.story.addObj(this);
@@ -457,6 +460,7 @@ class Board extends StoryObj {
             text: this.text,
             tags: this.tags,
             img: this.img,
+            audio: this.audio,
             btns: this.buttons.map(v => {
                 let o2 = {
                     l: v.label,
@@ -549,7 +553,8 @@ class User {
         await write("users/" + this.email + ".json", str, "utf8");
         if (!await access("projects/" + this.email)) {
             await mkdir("projects/" + this.email);
-            await mkdir("projects/" + this.email + "/images");
+            // await mkdir("projects/"+this.email+"/images");
+            // await mkdir("projects/"+this.email+"/audio");
         }
     }
     static from(data, socket) {
@@ -662,6 +667,7 @@ class Project {
     _data;
     story;
     images = [];
+    audio = [];
     getId() {
         return this.owner + "_" + this.name;
     }
@@ -799,6 +805,8 @@ async function getProject(email, pname) {
                 }
             }
         });
+        await mkdir(`projects/${email}/${pname}/images`);
+        await mkdir(`projects/${email}/${pname}/audio`);
     });
     if (!p) {
         console.log("...couldn't find file/folder");
@@ -806,7 +814,8 @@ async function getProject(email, pname) {
     }
     p.story = await Story.load(p);
     console.log("...found and loading!!!");
-    p.images = await readdir(`projects/${email}/${pname}/images`);
+    p.images = await readdir(`projects/${email}/${pname}/images`) || [];
+    p.audio = await readdir(`projects/${email}/${pname}/audio`) || [];
     projects.set(p.getId(), p);
     return p;
 }
@@ -1039,6 +1048,23 @@ io.on("connection", socket => {
         b.img = name;
         u.room().emit("setBGImage", u.email, id, name);
     });
+    socket.on("s_setBGAudio", async (id, name, f) => {
+        let u = await getWho(socket);
+        if (!u)
+            return;
+        let p = u.curProject;
+        if (!p)
+            return;
+        if (!p.audio.includes(name) && name != null) {
+            f(0);
+            return;
+        }
+        let b = p.story.getBoard(id);
+        if (!b)
+            return;
+        b.audio = name;
+        u.room().emit("setBGAudio", u.email, id, name);
+    });
     // socket.on("s_moveBoardsTo",async (list:{id:number,x:number,y:number}[])=>{
     //     let u = await getWho(socket);
     //     if(!u) return;
@@ -1100,6 +1126,47 @@ io.on("connection", socket => {
             p.images.push(name);
         }
         let url = "projects/" + p.owner + "/" + p.name + "/images/" + name;
+        await write(url, file);
+        f(url);
+    });
+    socket.on("s_addAudioFile", async (name, filedata, f) => {
+        let u = await getWho(socket);
+        if (!u)
+            return;
+        let p = u.curProject;
+        if (!p)
+            return;
+        // let base64Data = filedata.replace(/^data:image\/png;base64,/, "");
+        // let binaryData = Buffer.from(base64Data, 'base64');
+        // let url = "projects/"+p.owner+"/"+p.name+"/audio/"+name;
+        // await write(url,binaryData,"binary");
+        // console.log("going to write some audio data...");
+        let url = "";
+        f(url);
+    });
+    socket.on("s_getAudioFiles", async (f) => {
+        let u = await getWho(socket);
+        if (!u)
+            return;
+        let p = u.curProject;
+        if (!p)
+            return;
+        f(p.audio);
+    });
+    socket.on("s_uploadAudioFile", async (name, file, f) => {
+        if (name.includes("/"))
+            return;
+        let u = await getWho(socket);
+        if (!u)
+            return;
+        let p = u.curProject;
+        if (!p)
+            return;
+        console.log("$ start uploading audio file");
+        if (!p.audio.includes(name)) {
+            p.audio.push(name);
+        }
+        let url = "projects/" + p.owner + "/" + p.name + "/audio/" + name;
         await write(url, file);
         f(url);
     });
@@ -1196,12 +1263,12 @@ server.listen(3000, () => {
 });
 let rl = readline_1.default.createInterface(process.stdin, process.stdout);
 rl.on("line", (v) => {
-    v = v.substring(2);
+    // v = v.substring(2);
     if (!v)
         return;
     if (v == "stop") {
         process.exit(0);
     }
     // console.log("> "+v);
-    rl.write("> ");
+    // rl.write("> ");
 });