+++ src/play.ts
@@ -38,6 +38,8 @@ let playStory:Story;
 let allBoards:Board[] = [];
 
 let passage = document.querySelector(".passage");
+let imgCont = document.querySelector(".img-cont");
+let main = document.querySelector(".main");
 // let buttons = document.querySelector(".buttons");
 
 function madeChoice(btn:StoryButton,button:HTMLButtonElement){
@@ -49,19 +51,34 @@ function madeChoice(btn:StoryButton,button:HTMLButtonElement){
     button.classList.add("picked");
 }
 function scrollDown(){
-    scrollTo({left:0,top:document.body.scrollHeight,behavior:"smooth"});
+    main.scrollTo({left:0,top:document.body.scrollHeight,behavior:"smooth"});
 }
+
+let curImage:HTMLImageElement;
 async function loadBoard(b:Board){
     playData.locId = b._id;
     playData.save();
+    let imgUrl = b.img;
+
+    if(imgUrl){
+        if(curImage){
+            await wait(500);
+            document.body.classList.add("fade-out");
+            await wait(2000);
+            curImage.remove();
+            passage.textContent = "";
+            await wait(1000);
+            document.body.classList.remove("fade-out");
+        }
+    }
 
     let cont = document.createElement("div");
     passage.appendChild(cont);
 
-    let imgUrl = b.img;
-    console.log("img",imgUrl);
     if(imgUrl){
         let img = document.createElement("img");
+        img.draggable = false;
+        curImage = img;
         img.src = `${serverURL}/projects/${playStory.owner}/${playStory.filename}/images/${imgUrl}`;
         let success = false;
         await new Promise<void>(resolve=>{
@@ -74,13 +91,18 @@ async function loadBoard(b:Board){
             };
         });
         // if(success){
-            cont.appendChild(img);
-            await wait(4000);
+            // cont.appendChild(img);
+            imgCont.appendChild(img);
+            // document.body.insertAdjacentElement("beforebegin",img);
+            await wait(4100);
         // }
+
+        // wait for image anim
+        // await wait(10000);
     }
     
     let lines = b.text.replaceAll("\n"," ").split(". ").map(v=>v.trim());
-    await wait(500);
+    await wait(200); //500
     for(const l of lines){
         let div = document.createElement("div");
         div.className = "p-item";
@@ -98,7 +120,7 @@ async function loadBoard(b:Board){
         scrollDown();
         return;
     }
-    await wait(500);
+    await wait(300); //500
     
     let buttons = document.createElement("div");
     cont.appendChild(buttons);
@@ -120,7 +142,7 @@ async function loadBoard(b:Board){
         });
         
         scrollDown();
-        await wait(500);
+        await wait(300); //500
     }
 }
 
@@ -182,6 +204,7 @@ async function initPlay(){
     pdata.storyData.owner = pdata.owner;
     story = Story.load(pdata.storyData);
     console.log("story",story);
+    document.title = story.filename;
 
     //